pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-reports-databook"
version = "0.1.0"
description = "Annual report PDF to Excel databook extraction pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.10.0",
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "openpyxl>=3.1.0",
    "marker-pdf>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
databook-run = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
